refactor: consolidate duplicate exception package into exceptions; remove empty api.data package
- Move GlobalValidationExceptionHandler from ...exception (singular) to ...exceptions (plural),
  updating its package declaration. The ...exception package is removed.
- Move GlobalValidationExceptionHandlerScopeTest to match the new package location.
- Delete the empty, unreferenced src/main/.../api/data/Response.java (0 bytes) and its directory.
- Document in CHANGELOG.md (Breaking Changes) and MIGRATION.md (Package consolidation subsection).
